pyproject.toml
[project]
name = "terraform-provider-python"
version = "0.1.0"
description = "Terraform Provider Python"
authors = [{ name = "Ben Wart", email = "benwart@duck.com" }]
dependencies = [
  "grpcio>=1.62.2",
  "msgpack>=1.0.8",
  "anyio>=4.3.0",
  "loguru>=0.7.2",
  "grpcio-health-checking>=1.62.2",
  "cryptography>=42.0.5",
]
requires-python = "<3.13,>=3.12"
readme = "README.md"
license = { text = "MIT" }

[project.optional-dependencies]
dev = [
  "grpcio-tools>=1.62.2",
  "black>=24.4.2",
  "pytest>=8.1.2",
  "trio>=0.25.0",
  "pyinstaller>=6.6.0",
]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.black]
line-length = 160
target-version = ['py312']

[tool.ruff]
line-length = 160

[tool.pdm]
distribution = true
